requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
scipy>=1.10.0
Pillow>=9.5.0

# Data processing
albumentations>=1.3.0
opencv-python-headless>=4.8.0

# Evaluation
xtcocotools>=1.12

# Training utilities
tqdm>=4.65.0
PyYAML>=6.0

# Logging (optional)
wandb>=0.15.0

# Jupyter (for notebooks)
jupyter>=1.0.0
ipywidgets>=8.0.0
matplotlib>=3.7.0
